+++ src/main/java/com/comandante/creeper/player/BasicPlayerLevelStatsModifier.java
@@ -24,59 +24,59 @@ public class BasicPlayerLevelStatsModifier implements StatsModifier {
     private static double STRENGTH_CONSTANT_MODIFIER = 1.4;
     private static double MANA_CONSTANT_MODIFIER = 2;
 
-    public static int getMeleForLevel(int baseStat, int level) {
+    public static long getMeleForLevel(long baseStat, long level) {
         double v = (level) * sqrt(pow(level, MELE_CONSTANT_MODIFIER));
-        return (int) Math.floor(v) + baseStat;
+        return (long) Math.floor(v) + baseStat;
     }
 
-    public static int getWillpowerForLevel(int baseStat, int level) {
+    public static long getWillpowerForLevel(long baseStat, long level) {
         double v = (level) * sqrt(pow(level, WILLPOWER_CONSTANT_MODIFIER));
-        return (int) Math.floor(v) + baseStat;
+        return (long) Math.floor(v) + baseStat;
     }
 
-    public static int getAgileForLevel(int baseStat, int level) {
+    public static long getAgileForLevel(long baseStat, long level) {
         double v = (level) * sqrt(pow(level, AGILE_CONSTANT_MODIFIER));
-        return (int) Math.floor(v) + baseStat;
+        return (long) Math.floor(v) + baseStat;
     }
 
-    public static int getAimForLevel(int baseStat, int level) {
+    public static long getAimForLevel(long baseStat, long level) {
         double v = (level) * sqrt(pow(level, AIM_CONSTANT_MODIFIER));
-        return (int) Math.floor(v) + baseStat;
+        return (long) Math.floor(v) + baseStat;
     }
 
-    public static int getStrengthForLevel(int baseStat, int level) {
+    public static long getStrengthForLevel(long baseStat, long level) {
         double v = (level) * sqrt(pow(level, STRENGTH_CONSTANT_MODIFIER));
-        return (int) Math.floor(v) + baseStat;
+        return (long) Math.floor(v) + baseStat;
     }
 
-    public static int getArmorForLevel(int baseStat, int level) {
+    public static long getArmorForLevel(long baseStat, long level) {
         double v = (level) * sqrt(pow(level, ARMOR_CONSTANT_MODIFIER));
-        return (int) Math.floor(v) + baseStat;
+        return (long) Math.floor(v) + baseStat;
     }
 
-    public static int getHealthForLevel(int baseStat, int level) {
+    public static long getHealthForLevel(long baseStat, long level) {
         double v = (level) * sqrt(pow(level, HEALTH_CONSTANT_MODIFIER));
-        return (int) Math.floor(v) + baseStat;
+        return (long) Math.floor(v) + baseStat;
     }
 
-    public static int getManaForLevel(int baseStat, int level) {
+    public static long getManaForLevel(long baseStat, long level) {
         double v = (level) * sqrt(pow(level, MANA_CONSTANT_MODIFIER));
-        return (int) Math.floor(v) + baseStat;
+        return (long) Math.floor(v) + baseStat;
     }
 
     @Override
     public Stats modify(Player player) {
         PlayerMetadata playerMetadata = gameManager.getPlayerManager().getPlayerMetadata(player.getPlayerId());
         Stats baseStats = playerMetadata.getStats();
-        int level = Levels.getLevel(baseStats.getExperience());
-        int newMaxHealth = getHealthForLevel(baseStats.getMaxHealth(), level);
-        int newArmorRating = getArmorForLevel(baseStats.getArmorRating(), level);
-        int newStrengthRating = getStrengthForLevel(baseStats.getStrength(), level);
-        int newMaxMana = getManaForLevel(baseStats.getMaxMana(), level);
-        int newAimRating = getAimForLevel(baseStats.getAim(), level);
-        int newWillpowerRating = getWillpowerForLevel(baseStats.getWillpower(), level);
-        int newAgileRating = getAgileForLevel(baseStats.getAgile(), level);
-        int newMeleRating = getMeleForLevel(baseStats.getMeleSkill(), level);
+        long level = Levels.getLevel(baseStats.getExperience());
+        long newMaxHealth = getHealthForLevel(baseStats.getMaxHealth(), level);
+        long newArmorRating = getArmorForLevel(baseStats.getArmorRating(), level);
+        long newStrengthRating = getStrengthForLevel(baseStats.getStrength(), level);
+        long newMaxMana = getManaForLevel(baseStats.getMaxMana(), level);
+        long newAimRating = getAimForLevel(baseStats.getAim(), level);
+        long newWillpowerRating = getWillpowerForLevel(baseStats.getWillpower(), level);
+        long newAgileRating = getAgileForLevel(baseStats.getAgile(), level);
+        long newMeleRating = getMeleForLevel(baseStats.getMeleSkill(), level);
         StatsBuilder statsBuilder = new StatsBuilder(baseStats);
         statsBuilder.setMaxHealth(newMaxHealth);
         statsBuilder.setArmorRating(newArmorRating);
